+++ NEWS.md
@@ -1,3 +1,5 @@
 # r2dii.data 0.0.3.9000
 
+* The column `is_ultimate_owner` was removed from `ald_demo` (#22). That column was unused and unavailable to banks.
+
 * Added a `NEWS.md` file to track changes to the package.